requirement.txt
pandas
pyarrow
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
streamlit
requests
plotly